Split long dash app into multiple files
