+++ App/app_instance.py
@@ -1,0 +1,13 @@
+import dash
+import dash_bootstrap_components as dbc
+import dash_cytoscape as cyto
+
+app = dash.Dash(
+    __name__,
+    external_stylesheets=[dbc.themes.BOOTSTRAP],
+    suppress_callback_exceptions=True,
+)
+app.title = "MS2LDA Interactive Dashboard"
+
+# Include Cytoscape extra layouts
+cyto.load_extra_layouts()
